# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "SkoutWatch App OpenApi API Documentation",
        "description": "SkoutWatch App Using L5 Swagger OpenApi description",
        "contact": {
            "email": "schneidershades@gmail.com"
        },
        "version": "1.0"
    },
    "servers": [
        {
            "url": "http://skoutwatch.test",
            "description": "Local API server"
        },
        {
            "url": "https://dev-api.skoutwatch.com",
            "description": "Staging API server"
        },
        {
            "url": "https://staging.skoutwatch.com",
            "description": "Staging API server"
        },
        {
            "url": "https://api.skoutwatch.com",
            "description": "Live API server"
        }
    ],
    "paths": {
...
        "/api/v1/user/email/resend": {
            "post": {
                "tags": [
                    "OTP"
                ],
                "summary": "Verify email of a registered user",
                "description": "Verify email of a registered user",
                "operationId": "userResendVerifyEmailViaOTP",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/ResendVerificationEmailFormRequest"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Successful signin",
                        "content": {
                            "application/json": {}
                        }
                    },
                    "400": {
                        "description": "Bad Request"
                    },
                    "401": {
                        "description": "Unauthenticated"
                    },
                    "403": {
                        "description": "Forbidden"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/v1/user/email/resend-verify-Otp-with-link": {
            "post": {
                "tags": [
                    "OTP"
                ],
                "summary": "Verify email of a registered user",
                "description": "Verify email of a registered user",
                "operationId": "userResendVerifyEmailViaOTPWithLink",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/ResendVerificationEmailFormRequest"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Successful signin",
                        "content": {
                            "application/json": {}
                        }
                    },
                    "400": {
                        "description": "Bad Request"
                    },
                    "401": {
                        "description": "Unauthenticated"
                    },
                    "403": {
                        "description": "Forbidden"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/v1/user/email/verify": {
            "post": {
                "tags": [
                    "OTP"
                ],
                "summary": "Profile of a registered user",
                "description": "Profile of a registered user",
                "operationId": "userVerifyViaOTP",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/VerificationFormRequest"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Successful signin",
                        "content": {
                            "application/json": {}
                        }
                    },
                    "400": {
                        "description": "Bad Request"
                    },
                    "401": {
                        "description": "Unauthenticated"
                    },
                    "403": {
                        "description": "Forbidden"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/v1/verify/user": {
            "post": {
                "tags": [
                    "VerifyMe"
                ],
                "summary": "Verify registered user",
                "description": "Verify registered user",
                "operationId": "userResendVerifyMeAPI",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/StoreVerifyMeFormRequest"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Successful signin",
                        "content": {
                            "application/json": {}
                        }
                    },
                    "400": {
                        "description": "Bad Request"
                    },
                    "401": {
                        "description": "Unauthenticated"
                    },
                    "403": {
                        "description": "Forbidden"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/v1/verify/company": {
            "post": {
                "tags": [
                    "VerifyMe"
                ],
                "summary": "Verify registered company",
                "description": "Verify registered company",
                "operationId": "companyResendVerifyMeAPI",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/StoreVerifyCompanyFormRequest"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Successful signin",
                        "content": {
                            "application/json": {}
                        }
                    },
                    "400": {
                        "description": "Bad Request"
                    },
                    "401": {
                        "description": "Unauthenticated"
                    },
                    "403": {
                        "description": "Forbidden"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        }
    },
    "components": {
        "schemas": {
...
            "ResendVerificationEmailFormRequest": {
                "title": "User Resend OTP Form Request Fields",
                "description": "User Resend OTP request body data",
                "required": [
                    "first_name"
                ],
                "properties": {
                    "email": {
                        "title": "User reset email",
                        "description": "Reset email of the user",
                        "type": "string",
                        "example": "info@tonote.com"
                    }
                },
                "type": "object"
            },
...
            "VerificationFormRequest": {
                "title": "User Verify OTP Form Request Fields",
                "description": "User Verify OTP request body data",
                "required": [
                    "first_name"
                ],
                "properties": {
                    "email": {
                        "title": "User reset email",
                        "description": "Reset email of the user",
                        "type": "string",
                        "example": "info@tonote.com"
                    },
                    "otp": {
                        "title": "User OTP",
                        "description": "User OTP",
                        "type": "string",
                        "example": "56777"
                    },
                    "channel": {
                        "title": "User channel",
                        "description": "User channel",
                        "type": "string",
                        "example": "document"
                    }
                },
                "type": "object"
            },
...
            "StoreVerifyCompanyFormRequest": {
                "title": "Company Verify Me Form Request Fields",
                "description": "Company Verify Me request body data",
                "required": [
                    "first_name"
                ],
                "properties": {
                    "registration_company_number": {
                        "title": "Company reset rc",
                        "description": "Reset rc of the Company",
                        "type": "string",
                        "example": "RC293332"
                    },
                    "type": {
                        "title": "Company type rc",
                        "description": "type rc of the Company",
                        "type": "string",
                        "example": "Business|Limited Company|Incorprated Trustee"
                    }
                },
                "type": "object"
            },
            "StoreVerifyMeFormRequest": {
                "title": "User Verify Me Form Request Fields",
                "description": "User Verify Me request body data",
                "required": [
                    "first_name"
                ],
                "properties": {
                    "type": {
                        "title": "Verify type",
                        "description": "Verify type",
                        "type": "string",
                        "example": "bvn,nin,drivers_license,vnin"
                    },
                    "value": {
                        "title": "verify value",
                        "description": "verify value",
                        "type": "string",
                        "example": "1029333022"
                    }
                },
                "type": "object"
            }
        },
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "scheme": "bearer"
            }
        }
    },
...
}